Implement generic logger_mocked
current_logger() did not have a united implementation so far.
This commit implements a generic, fully configurable mocked actor for
use in all repository unit tests. All old implementations are removed.
