+++ repos/system_upgrade/el7toel8/actors/kernel/checkinstalledkernels/tests/unit_test.py
@@ -3,22 +3,11 @@ from collections import namedtuple
 from leapp import reporting
 from leapp.libraries.actor import library
 from leapp.libraries.common.config import architecture
-from leapp.libraries.common.testutils import create_report_mocked, CurrentActorMocked
+from leapp.libraries.common.testutils import create_report_mocked, CurrentActorMocked, logger_mocked
 from leapp.libraries.stdlib import api
 from leapp.models import InstalledRedHatSignedRPM, RPM
 
 RH_PACKAGER = 'Red Hat, Inc. <http://bugzilla.redhat.com/bugzilla>'
-
-
-class mocked_logger(object):
-    def __init__(self):
-        self.errmsg = None
-
-    def error(self, *args):
-        self.errmsg = args
-
-    def __call__(self):
-        return self
 
 
 def mocked_consume(pkgs):  # pkgs = [(name, version-number)]
@@ -42,7 +31,7 @@ s390x_pkgs_multi = [('kernel', 957), ('something', 957), ('kernel', 956)]
 
 def test_single_kernel_s390x(monkeypatch):
     monkeypatch.setattr(api, 'current_actor', CurrentActorMocked())
-    monkeypatch.setattr(api, 'current_logger', mocked_logger())
+    monkeypatch.setattr(api, 'current_logger', logger_mocked())
     monkeypatch.setattr(api, 'consume', mocked_consume(s390x_pkgs_single))
     monkeypatch.setattr(reporting, 'create_report', create_report_mocked())
     library.process()
@@ -56,7 +45,7 @@ def test_single_kernel_s390x(monkeypatch):
 
 def test_multi_kernel_s390x(monkeypatch):
     monkeypatch.setattr(api, 'current_actor', CurrentActorMocked())
-    monkeypatch.setattr(api, 'current_logger', mocked_logger())
+    monkeypatch.setattr(api, 'current_logger', logger_mocked())
     monkeypatch.setattr(api, 'consume', mocked_consume(s390x_pkgs_multi))
     monkeypatch.setattr(reporting, 'create_report', create_report_mocked())
     library.process()
@@ -74,14 +63,14 @@ versioned_kernel_pkgs = [('kernel', 456), ('kernel', 789), ('kernel', 1234)]
 
 def test_newest_kernel(monkeypatch):
     monkeypatch.setattr(api, 'current_actor', CurrentActorMocked(kernel='3.10.0-1234.21.1.el7.x86_64'))
-    monkeypatch.setattr(api, 'current_logger', mocked_logger())
+    monkeypatch.setattr(api, 'current_logger', logger_mocked())
     monkeypatch.setattr(api, 'consume', mocked_consume(versioned_kernel_pkgs))
     monkeypatch.setattr(reporting, 'create_report', create_report_mocked())
     library.process()
     assert not reporting.create_report.called
 
     monkeypatch.setattr(api, 'current_actor', CurrentActorMocked(kernel='3.10.0-456.43.1.el7.x86_64'))
-    monkeypatch.setattr(api, 'current_logger', mocked_logger())
+    monkeypatch.setattr(api, 'current_logger', logger_mocked())
     monkeypatch.setattr(api, 'consume', mocked_consume(versioned_kernel_pkgs))
     monkeypatch.setattr(reporting, 'create_report', create_report_mocked())
     library.process()
@@ -89,7 +78,7 @@ def test_newest_kernel(monkeypatch):
     assert reporting.create_report.report_fields['title'] == 'Newest installed kernel not in use'
 
     monkeypatch.setattr(api, 'current_actor', CurrentActorMocked(kernel='3.10.0-789.35.2.el7.x86_64'))
-    monkeypatch.setattr(api, 'current_logger', mocked_logger())
+    monkeypatch.setattr(api, 'current_logger', logger_mocked())
     monkeypatch.setattr(api, 'consume', mocked_consume(versioned_kernel_pkgs))
     monkeypatch.setattr(reporting, 'create_report', create_report_mocked())
     library.process()
